:sparkles: refactor into GraphWatcher class
